modified the bytes array for OpenMP version

+++ src/openmp/stream_openmp.c
@@ -190,15 +190,15 @@ int main(int argc, char *argv[])
 		3 * sizeof(STREAM_TYPE) * stream_array_size, // Add
 		3 * sizeof(STREAM_TYPE) * stream_array_size, // Triad
 		// Gather Kernels
-		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(int))) * stream_array_size), // GATHER copy
-		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(int))) * stream_array_size), // GATHER Scale
-		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(int))) * stream_array_size), // GATHER Add
-		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(int))) * stream_array_size), // GATHER Triad
+		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(ssize_t))) * stream_array_size), // GATHER copy
+		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(ssize_t))) * stream_array_size), // GATHER Scale
+		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(ssize_t))) * stream_array_size), // GATHER Add
+		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(ssize_t))) * stream_array_size), // GATHER Triad
 		// Scatter Kernels
-		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(int))) * stream_array_size), // SCATTER copy
-		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(int))) * stream_array_size), // SCATTER Scale
-		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(int))) * stream_array_size), // SCATTER Add
-		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(int))) * stream_array_size), // SCATTER Triad
+		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(ssize_t))) * stream_array_size), // SCATTER copy
+		(((2 * sizeof(STREAM_TYPE)) + (1 * sizeof(ssize_t))) * stream_array_size), // SCATTER Scale
+		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(ssize_t))) * stream_array_size), // SCATTER Add
+		(((3 * sizeof(STREAM_TYPE)) + (2 * sizeof(ssize_t))) * stream_array_size), // SCATTER Triad
 	};
 
 	double   flops[NUM_KERNELS] = {
